# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Customer API Documentation",
        "description": "Documentation de l'API Customer",
        "contact": {
            "email": "support@example.com"
        },
        "version": "2.0.0"
    },
    "paths": {
...
        "/techtablet-sellers": {
            "get": {
                "tags": [
                    "TechtabletSellers"
                ],
                "summary": "Liste tous les vendeurs Techtablet",
                "operationId": "1b5d8df36d9b3badcda19da995df33bd",
                "responses": {
                    "200": {
                        "description": "Liste des vendeurs Techtablet récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TechtabletSeller"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des vendeurs Techtablet récupérée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "TechtabletSellers"
                ],
                "summary": "Crée un nouveau vendeur Techtablet",
                "operationId": "7443a4d718baccc2c7da92da0ae982a6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTechtabletSellerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vendeur Techtablet créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TechtabletSeller"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vendeur Techtablet créé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/techtablet-sellers/{id}": {
            "get": {
                "tags": [
                    "TechtabletSellers"
                ],
                "summary": "Affiche les détails d'un vendeur Techtablet spécifique",
                "operationId": "7d2e887c1e156edeb12d2f93638ad736",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du vendeur Techtablet",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vendeur Techtablet récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TechtabletSeller"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vendeur Techtablet récupéré avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vendeur Techtablet non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "TechtabletSellers"
                ],
                "summary": "Met à jour un vendeur Techtablet existant",
                "operationId": "26d84c4c3f5925cc51388fd1d7efeed6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du vendeur Techtablet à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTechtabletSellerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vendeur Techtablet mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TechtabletSeller"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vendeur Techtablet mis à jour avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vendeur Techtablet non trouvé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "TechtabletSellers"
                ],
                "summary": "Supprime un vendeur Techtablet",
                "operationId": "8506d5f603d13b59404134f6a03158e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du vendeur Techtablet à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vendeur Techtablet supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vendeur Techtablet supprimé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vendeur Techtablet non trouvé"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreTechtabletSellerRequest": {
                "required": [
                    "first_name",
                    "employee_code"
                ],
                "properties": {
                    "first_name": {
                        "description": "Prénom",
                        "type": "string",
                        "maxLength": 50,
                        "example": "Pierre"
                    },
                    "last_name": {
                        "description": "Nom",
                        "type": "string",
                        "maxLength": 50,
                        "example": "Martin",
                        "nullable": true
                    },
                    "primary_phone": {
                        "description": "Téléphone principal",
                        "type": "string",
                        "maxLength": 20,
                        "example": "+33123456789",
                        "nullable": true
                    },
                    "secondary_phone": {
                        "description": "Téléphone secondaire",
                        "type": "string",
                        "maxLength": 20,
                        "example": "+33612345678",
                        "nullable": true
                    },
                    "email": {
                        "description": "Email professionnel",
                        "type": "string",
                        "maxLength": 150,
                        "example": "pierre.martin@techtablet.com",
                        "nullable": true
                    },
                    "job_title": {
                        "description": "Poste/emploi occupé",
                        "type": "string",
                        "maxLength": 100,
                        "example": "Commercial Senior",
                        "nullable": true
                    },
                    "employee_code": {
                        "description": "Code employé unique",
                        "type": "string",
                        "maxLength": 30,
                        "example": "EMP001"
                    },
                    "digital_signature": {
                        "description": "Signature numérique",
                        "type": "string",
                        "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA...",
                        "nullable": true
                    },
                    "is_active": {
                        "description": "Statut actif",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "UpdateTechtabletSellerRequest": {
                "properties": {
                    "first_name": {
                        "description": "Prénom",
                        "type": "string",
                        "maxLength": 50,
                        "example": "Pierre"
                    },
                    "last_name": {
                        "description": "Nom",
                        "type": "string",
                        "maxLength": 50,
                        "example": "Martin",
                        "nullable": true
                    },
                    "primary_phone": {
                        "description": "Téléphone principal",
                        "type": "string",
                        "maxLength": 20,
                        "example": "+33123456789",
                        "nullable": true
                    },
                    "secondary_phone": {
                        "description": "Téléphone secondaire",
                        "type": "string",
                        "maxLength": 20,
                        "example": "+33612345678",
                        "nullable": true
                    },
                    "email": {
                        "description": "Email professionnel",
                        "type": "string",
                        "maxLength": 150,
                        "example": "pierre.martin@techtablet.com",
                        "nullable": true
                    },
                    "job_title": {
                        "description": "Poste/emploi occupé",
                        "type": "string",
                        "maxLength": 100,
                        "example": "Commercial Senior",
                        "nullable": true
                    },
                    "employee_code": {
                        "description": "Code employé unique",
                        "type": "string",
                        "maxLength": 30,
                        "example": "EMP001"
                    },
                    "digital_signature": {
                        "description": "Signature numérique",
                        "type": "string",
                        "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA...",
                        "nullable": true
                    },
                    "is_active": {
                        "description": "Statut actif",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "TechtabletSeller": {
                "title": "TechtabletSeller",
                "description": "Modèle de vendeur Techtablet",
                "properties": {
                    "id_techtablet_seller": {
                        "description": "ID unique du vendeur Techtablet",
                        "type": "integer",
                        "example": 1
                    },
                    "first_name": {
                        "description": "Prénom",
                        "type": "string",
                        "maxLength": 50,
                        "example": "Pierre"
                    },
                    "last_name": {
                        "description": "Nom",
                        "type": "string",
                        "maxLength": 50,
                        "example": "Martin",
                        "nullable": true
                    },
                    "primary_phone": {
                        "description": "Téléphone principal",
                        "type": "string",
                        "maxLength": 20,
                        "example": "+33123456789",
                        "nullable": true
                    },
                    "secondary_phone": {
                        "description": "Téléphone secondaire",
                        "type": "string",
                        "maxLength": 20,
                        "example": "+33612345678",
                        "nullable": true
                    },
                    "email": {
                        "description": "Email professionnel",
                        "type": "string",
                        "maxLength": 150,
                        "example": "pierre.martin@techtablet.com",
                        "nullable": true
                    },
                    "job_title": {
                        "description": "Poste/emploi occupé",
                        "type": "string",
                        "maxLength": 100,
                        "example": "Commercial Senior",
                        "nullable": true
                    },
                    "employee_code": {
                        "description": "Code employé unique",
                        "type": "string",
                        "maxLength": 30,
                        "example": "EMP001"
                    },
                    "digital_signature": {
                        "description": "Signature numérique",
                        "type": "string",
                        "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA...",
                        "nullable": true
                    },
                    "is_active": {
                        "description": "Statut actif",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de dernière modification",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}